% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-real.R
\name{all_real}
\alias{all_real}
\title{all_real}
\usage{
all_real(x, include_na = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{vector.}

\item{include_na}{Logical if `NA` should also be considered as non-real.}

\item{verbose}{Logical if number of non-real numbers should be printed as message.}
}
\value{
vector
}
\description{
All real
}
\details{
The function checks if a vector contains \code{NaN} or \code{Inf} values. Input
must be a vector. If \code{include_na = TRUE} also \code{NA} values are considered.
}
\examples{
x <- c(1, 2, 3, NA, 5, NaN, Inf)
all_real(x)

y <- c(1, 2, 3, 4, 5, 6, 7)
all_real(y)

}
