% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_table.R
\name{as_data_table}
\alias{as_data_table}
\title{as_data_table}
\usage{
as_data_table(x, na.rm = TRUE, return_df = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{RasterLayer.}

\item{na.rm}{Logical if `NA` should be removed.}

\item{return_df}{If TRUE, `data.frame` instead of `data.table` is returned.}

\item{verbose}{If TRUE, progress report is printed.}
}
\value{
data.table
}
\description{
Convert raster object to data.table
}
\details{
Function to convert a RasterLayer into a data.table. For rasters that cannot be
hold in memory, values are written blockwise.
}
\references{
Adapted from Etienne B. Racine (https://gist.github.com/etiennebr/9515738)
}
