% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-temp.R
\name{clean_temp}
\alias{clean_temp}
\title{Delete temporary files}
\usage{
clean_temp(
  base_dir = dirname(tempdir()),
  minutes_keep = 0,
  list_files = FALSE,
  delete = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{base_dir}{Character string with the path to the temporary directory.}

\item{minutes_keep}{Numeric. Age in minutes from which on folders are deleted; more recent folders are kept.}

\item{list_files}{Logical, whether all files should be listed.}

\item{delete}{Logical, weather all files should be deleted.}

\item{verbose}{Logical, whether messages are sent to the console.}
}
\value{
(invisible) list
}
\description{
Delete temporary files
}
\details{
The function deletes all files in the temporary folder which are older than \code{minutes_keep}.
}
\examples{
\dontrun{
clean_temp()
}

}
\author{
Code based on scripts by Markus Samek
}
