% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid_unique.R
\name{expand_grid_unique}
\alias{expand_grid_unique}
\title{expand_grid_unique}
\usage{
expand_grid_unique(x, y, equals = FALSE)
}
\arguments{
\item{x}{vector x.}

\item{y}{vector y}

\item{equals}{If TRUE also same combinations are included}
}
\value{
vector
}
\description{
Expand grid unique
}
\details{
The function returns all possible combinations of two vectors.
}
\examples{
x <- 1:5
y <- 1:5
expand_grid_unique(x, y)

}
