% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_real.R
\name{is_real}
\alias{is_real}
\title{is_real}
\usage{
is_real(x, include_na = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{vector.}

\item{include_na}{Logical if `NA` should also be considered as non-real.}

\item{verbose}{Logical if number of non-real numbers should be printed as message.}
}
\value{
vector
}
\description{
Checks if vector includes real numbers, i.e. not `NaN` or `Inf`.
}
\details{
The function checks if a vector contains real numbers and returns `FALSE` for
`NaN` or `Inf` values. Input must be a vector. If `include_na = TRUE` also `NA`s are considered.
}
\examples{
x <- c(1, 2, 3, NA, 5, NaN, Inf)
is_real(x)
is_real(x, include_na = TRUE)

y <- c(1, 2, 3, 4, 5, 6, 7)
is_real(y)

}
