% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{load_packages}
\usage{
load_packages(packages, install = "ask", verbose = TRUE)
}
\arguments{
\item{packages}{String with package names}

\item{install}{If "yes" missing packages are automatically installed. If "no" only already
installed packages are loaded. If "ask", an user input is needed.}

\item{verbose}{Print messages}
}
\value{
NA
}
\description{
Load pacakges
}
\details{
The function loads libraries. It's possible to provide a vector with string.
Missing packages can be installed automatically
}
\examples{
\dontrun{
load_packages(packages = c("dplyr", "spatstat"))
}

}
