% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_my_dataset.R
\docType{data}
\name{my_dataset}
\alias{my_dataset}
\title{mon jeu de données}
\format{
data.frame
}
\source{
\url{google.com}{unlien}
}
\usage{
my_dataset
}
\description{
Un jeu de données test
}
\author{
Colin \email{colin@thinkr.fr}
}
\keyword{data}
