% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Envelope.R
\name{Plot.Envelope}
\alias{Plot.Envelope}
\title{Function to plot simulation envelopes}
\usage{
Plot.Envelope(input, labels = c("clustering", " randomness", "segregation"),
  title = NULL, xlab = NULL, ylab = NULL, full_fun = T,
  standarized = F)
}
\arguments{
\item{input}{[\code{envelope(1)}]\cr Envelope object of the spatstat package}

\item{labels}{[\code{string(3)}]\cr Name of the labels
\cr 1 = obs > hi
\cr 2 = lo < obs < hi
\cr 3 = obs < lo}

\item{title}{[\code{string(1)}]\cr Title of the plot}

\item{xlab}{[\code{string(1)}]\cr x label of the plot}

\item{ylab}{[\code{string(1)}]\cr y label of the plot}

\item{full_fun}{[\code{logical(1)}]\cr Plot full function or only line to indicate}

\item{standarized}{[\code{logical(1)}]\cr If TRUE obs = obs - theo
deviation from null model}
}
\description{
This functions provides a plotting style for envelope objects of the spatstat
package (for more information please see ?spatstat::envelope). The location of the
observed value in relation to the simulation envelopes of the null model data is
indicated by an additional color bar at the bottom of the plot.
}
