% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_calculate_dist}
\alias{rcpp_calculate_dist}
\title{rcpp_calculate_dist}
\usage{
rcpp_calculate_dist(matrix, max_dist)
}
\arguments{
\item{matrix}{Matrix with x and y coordinates}

\item{max_dist}{Maximum distance to consider}
}
\value{
vector
}
\description{
Calculate sum of distances to neighbours
}
\details{
\code{Rcpp} implementation
}
\seealso{
\code{\link{dist}}
}
