% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_depends.R
\name{remove_depends}
\alias{remove_depends}
\title{remove_depends}
\usage{
remove_depends(pkg, recursive = FALSE)
}
\arguments{
\item{pkg}{R object to serialize.}

\item{recursive}{Name of the file where the R object is saved (file extension must be .rds).}
}
\description{
Remove dependencies
}
\details{
Function removes a package with all dependencies
}
\examples{
\dontrun{
remove_depends("dplyr")
}

}
