% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tmp.R
\name{remove_tmp}
\alias{remove_tmp}
\title{remove_tmp}
\usage{
remove_tmp(delete = FALSE)
}
\arguments{
\item{delete}{Logical if folders should be deleted}
}
\value{
vector
}
\description{
Remove temporary folders
}
\details{
The functions searches for all temporary folders and returns them as a vector.
Can also delete all files using `unlink(x, recursive = TRUE, force = TRUE, expand = TRUE)`
}
\examples{
remove_tmp()

}
