% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_infinite.R
\name{replace_infinite}
\alias{replace_infinite}
\title{replace_infinite}
\usage{
replace_infinite(x, what = c("NaN", "Inf"), value = c(NA, NA),
  verbose = TRUE)
}
\arguments{
\item{x}{vector.}

\item{what}{What infinite values should be replaced. See details.}

\item{value}{Value that will be used to replace selected infinite values.}

\item{verbose}{Print warning messages.}
}
\value{
vector
}
\description{
Replace infinite
}
\details{
The function converts all `NaN` and `Inf` values to NA. Input must be a vector.
What can be either `what = c("NaN", "Inf")` to replace both `NaN` and `Inf` or just
one of the two, e.g. `what = "NaN`.
}
\examples{
vec <- c(1, 2, 3, NA, 5, NaN, Inf)
replace_infinite(vec)
replace_infinite(vec, what = "NaN", value = NA)
replace_infinite(vec, value = c(NA, 0))

}
