% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rslurm_missing.R
\name{rslurm_missing}
\alias{rslurm_missing}
\title{rslurm_missing}
\usage{
rslurm_missing(x)
}
\arguments{
\item{x}{slurm_job object.}
}
\value{
vector
}
\description{
rslurm missing
}
\details{
The function checks if all .RDS files are present in the \code{rslurm_job} project.
For more information, please see the \code{rslurm} package.
}
\examples{
\dontrun{
rslurm_missing(x = sbatch_results)
}

}
