% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rslurm_read.R
\name{rslurm_read}
\alias{rslurm_read}
\title{rslurm_read}
\usage{
rslurm_read(x, pattern, as_list = TRUE)
}
\arguments{
\item{x}{slurm_job object.}

\item{pattern}{String with pattern to match.}

\item{as_list}{Logical if result should be returned as \code{list} or \code{data.frame}.}
}
\value{
list
}
\description{
rslurm read
}
\details{
This functions reads all \code{.rds} files matching the \code{pattern} argument.
This can be useful if results were not saved using the default templates of the
\code{rslurm} package. For more information, please see the \code{rslurm} package.
}
\examples{
\dontrun{
rslurm_read(x = sbatch_results)
}

}
