% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{runShiny}
\alias{runShiny}
\title{Launch interactive shiny application}
\usage{
runShiny(...)
}
\arguments{
\item{...}{Arguments are passed to \code{\link{runApp}}.}
}
\value{
This function does not return anything
}
\description{
Launches interactive shiny session in the browser.
}
\examples{
\dontrun{
library(support5860)
runShiny()
}
}
\seealso{
\code{\link{runApp}}
}
