% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_Function_ggplot.R
\name{Save.Function.ggplot}
\alias{Save.Function.ggplot}
\title{Function to save ggplot}
\usage{
Save.Function.ggplot(object, file, ...)
}
\arguments{
\item{object}{[\code{R object(1)}]\cr R object to save}

\item{file}{[\code{string(1)}]\cr Path and name where the file is saved}
}
\description{
This function checks if the file aready exists. See ?ggplot2::ggsave() for more information
}
