% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_Function_rds.R
\name{Save.Function.rds}
\alias{Save.Function.rds}
\title{Function to save rds objects}
\usage{
Save.Function.rds(file = stop("'file' must be specified"), ...)
}
\description{
This function checks if the file aready exists. See ?base::saveRDS() for more information
}
