% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-ggplot.R
\name{save_ggplot}
\alias{save_ggplot}
\title{save_ggplot}
\usage{
save_ggplot(
  plot,
  filename = NULL,
  path = NULL,
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{plot}{Plot to save.}

\item{filename}{File name to create on disk. The device (e.g. jpeg) must be specified as file extension.}

\item{path}{Path to save plot to (combined with filename).}

\item{overwrite}{If true, existing file is overwritten.}

\item{verbose}{If true, messages are printed to the console.}

\item{...}{Options passed to ggsave.}
}
\description{
Save ggplot
}
\details{
This function checks if the plot aready exists before saving it. If the file aready
exists and \code{overwrite = FALSE} (default), the file is not saved. The filename and path
is combined using \code{file.path()}. See \code{?ggplot2::ggsave()} for more information.

It is possible to set \code{path = ""} and use the entire path including the file
name as \code{filename} argument.
}
\examples{
\dontrun{
x <- runif(n = 100)
y <- runif(n = 100)
result_plot <- ggplot2::ggplot() + ggplot2::geom_point(ggplot2::aes(x = x, y = y))
save_ggplot(result_plot, filename = "result_ggpplot.jpeg")
}

}
\seealso{
\code{\link[ggplot2]{ggsave}}
}
