% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Path.R
\name{Set.Path}
\alias{Set.Path}
\title{Set path to working directory}
\usage{
Set.Path()
}
\value{
[\code{String(1)}]\cr WD for Hainich scripts
}
\description{
This function automatically detects the operating system and sets the path
}
