% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_environment.R
\name{show_environment}
\alias{show_environment}
\title{show_environment}
\usage{
show_environment(
  what = NULL,
  sort = "size",
  units = "Mb",
  decreasing = TRUE,
  n = NULL
)
}
\arguments{
\item{what}{List with objects to show object size. If \code{NULL} all objects in environment are used.}

\item{sort}{Sort the resulting dataframe either by "size", "name".}

\item{units}{Unit used to format memory usage. For more details, see \code{?object.size}.}

\item{decreasing}{Logical if results should be sorted in decreasing order.}

\item{n}{Number of top rows to print.}
}
\value{
data.frame
}
\description{
Show memory use
}
\details{
Functions returns the name, class and size of all objects in the environment. The
resulting data frame can be sorted by different arguments.
}
\examples{
vec <- 1:100
vec_2 <- runif(n = 10)
df <- data.frame(1:5000)
mat <- matrix(c(1:100), ncol = 20, nrow = 5)
list <- list(x = vec, y = df, z = mat)

show_environment()

}
\seealso{
\code{\link{object.size}}
}
