% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_temp.R
\name{show_temp}
\alias{show_temp}
\title{show_temp}
\usage{
show_temp(remove = FALSE, verbose = TRUE)
}
\arguments{
\item{remove}{Logical if folders should be deleted.}

\item{verbose}{Logical if messages should be printed.}
}
\value{
vector
}
\description{
Remove temporary folders
}
\details{
The functions searches for all temporary folders and returns them as a vector.
Can also delete all files using `unlink(x, recursive = TRUE, force = TRUE, expand = TRUE)`
}
\examples{
show_temp()

}
