% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_to_cluster.R
\name{submit_to_cluster}
\alias{submit_to_cluster}
\title{submit_to_cluster}
\usage{
submit_to_cluster(...)
}
\arguments{
\item{...}{Parameters passed on to \code{clustermq::Q()}.}
}
\value{
list
}
\description{
Submit to cluster
}
\details{
Wrapper around the function \code{Q()} of the *clustermq* package. The submitting and
finishing time are printed
}
\examples{
fx = function(x) x * 2
submit_to_cluster(fx, x = 1:3, n_jobs = 1)
list(2,4,6)


}
\seealso{
\code{\link{Q}}
}
