% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.CCA-package.R
\docType{package}
\name{Support.CCA-package}
\alias{Support.CCA}
\alias{Support.CCA-package}
\title{Support.CCA: SUPPORT RECOVERY OF CANONICAL DIRECTIONS IN SPARSE CCA}
\description{
Estimates the support of canonical direction vectors in case of
             sparse canonical correlation analysis.
}
\author{
\strong{Maintainer}: Nilanjana Laha \email{nlaha@hsph.harvard.edu}

Authors:
\itemize{
  \item Rajarshi Mukherjee \email{ram521@mail.harvard.edu}
}

}
\keyword{internal}
