% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\docType{package}
\name{UtilityFunctions-package}
\alias{UtilityFunctions}
\alias{UtilityFunctions-package}
\title{UtilityFunctions}
\description{
\emph{UtilityFunctions} UtilityFunctions ,,,
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mhesselbarth/UtilityFunctions}
  \item Report bugs at \url{https://github.com/mhesselbarth/UtilityFunctions/issues}
}

}
\author{
\strong{Maintainer}: Maximilian H.K. Hesselbarth \email{maximililan.hesselbarth@uni-goettingen.de} (0000-0003-1125-9918)

}
