% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero-linked-pdf.R
\name{zotero_linked_files}
\alias{zotero_linked_files}
\title{zotero_linked_files}
\usage{
zotero_linked_files(lib_file = file.choose(), full_name = TRUE)
}
\arguments{
\item{lib_file}{A string with the path to the Zotero library exported as a CSV file.}

\item{full_name}{Logical if full path to the files or only the file names should be returned}
}
\value{
vector
}
\description{
List linked files
}
\details{
This function reads a CSV file exported from Zotero and extracts the information about
the files linked to the references in the library.

To export your library from Zotero, go to the menu `File > Export Library...`
and choose the CSV format.
}
\examples{
\dontrun{zotero_linked_files()}

}
\references{
Adapted from Daniel Vartanian (https://gist.github.com/danielvartan/924817b7e4b69212beb217f339c37a3f)
}
\keyword{internal}
