% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-temp.R
\name{clean_temp}
\alias{clean_temp}
\title{Delete temporary files}
\usage{
clean_temp(
  base_dir = dirname(tempdir()),
  minutes_keep = 0,
  list_files = FALSE,
  delete = FALSE,
  verbose = getOption("energyTools.verbose", TRUE)
)
}
\arguments{
\item{base_dir}{Character with path to temp files.}

\item{minutes_keep}{Numeric with threshold in minutes for files to keep.}

\item{list_files}{Logical if all files should be listed.}

\item{delete}{Logical if all files should be deleted.}

\item{verbose}{Logical if messages are send to console.}
}
\value{
(invisible) list
}
\description{
Function to delete temporary files
}
\details{
The function deletes all files in the temporary folder which are older than \code{minutes_keep}.
}
\examples{
\dontrun{
clean_temp()
}

}
\author{
Code based on scripts of Markus Samek
}
