% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Envelope.R
\name{Plot.Envelope}
\alias{Plot.Envelope}
\title{Function to plot simulation envelopes}
\usage{
Plot.Envelope(input, labels = c("clustering", " randomness", "segregation"))
}
\arguments{
\item{input}{[\code{envelope(1)}]\cr Envelope object of the spatstat package}

\item{labels}{[\code{string(3)}]\cr Name of the labels.
\cr 1 = obs > hi
\cr 2 = lo < obs < hi
\cr 3 = obs < lo}
}
\description{
This functions provides a plotting style for envelope objects of the spatstat
package (for more information please see ?spatstat::envelope)
}
