% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_Function_ggplot.R
\name{Save.Function.ggplot}
\alias{Save.Function.ggplot}
\title{Function to save ggplot}
\usage{
Save.Function.ggplot(file = stop("'file' must be specified"), ...)
}
\description{
This function checks if the file aready exists. See ?ggplot2::ggsave() for more information
}
