% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_Function_ggplot.R
\name{Save.Function.ggplot}
\alias{Save.Function.ggplot}
\title{Function to save ggplot}
\usage{
Save.Function.ggplot(plot, filename, path, ...)
}
\arguments{
\item{plot}{[\code{ggplot(1)}]\cr Plot to save}

\item{filename}{[\code{string(1)}]\cr File name to create on disk}
}
\description{
This function checks if the plot aready exists before saving it.
See ?ggplot2::ggsave() for more information
}
