% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_Function_rds.R
\name{Save.Function.rds}
\alias{Save.Function.rds}
\title{Function to save rds objects}
\usage{
Save.Function.rds(object, filename, path = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{object}{[\code{R object(1)}]\cr R object to serialize}

\item{file}{[\code{string(1)}]\cr Connection or the name of the file where the R object is saved}
}
\description{
This function checks if the file aready exists before saving it.
See ?base::saveRDS() for more information
}
