% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ggplot.R
\name{save_ggplot}
\alias{save_ggplot}
\title{save_ggplot}
\usage{
save_ggplot(plot, filename = NULL, path = NULL, overwrite = FALSE,
  ...)
}
\arguments{
\item{plot}{Plot to save}

\item{filename}{File name to create on disk. The device (e.g. jpeg) must be specified as file extension}

\item{path}{Path to save plot to (combined with filename)}

\item{overwrite}{Overwrite if the plot alreay exists}

\item{...}{Options passed to ggsave}
}
\description{
save_ggplot
}
\details{
This function checks if the plot aready exists before saving it.
See ?ggplot2::ggsave() for more information.
}
