% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ggplot.R
\name{save_ggplot}
\alias{save_ggplot}
\title{save_ggplot}
\usage{
save_ggplot(plot, filename = NULL, path = NULL, overwrite = FALSE,
  rm_plot = FALSE, ...)
}
\arguments{
\item{plot}{Plot to save.}

\item{filename}{File name to create on disk. The device (e.g. jpeg) must be specified as file extension.}

\item{path}{Path to save plot to (combined with filename).}

\item{overwrite}{If true, existing file is overwritten.}

\item{rm_plot}{If true, the plot is removed form the environment.}

\item{...}{Options passed to ggsave.}
}
\description{
Save ggplot
}
\details{
This function checks if the plot aready exists before saving it. If the file aready
exists and `overwrite = FALSE` (default), the file is not saved. The filename and path
is combined using `file.path()`. See `?ggplot2::ggsave()` for more information.
}
\examples{
\dontrun{
x <- runif(n = 100)
y <- runif(n = 100)
result_plot <- ggplot2::ggplot() + ggplot2::geom_point(ggplot2::aes(x = x, y = y))
save_ggplot(result_plot, filename = "result_ggpplot.jpeg")
}

}
\seealso{
\code{\link{ggsave}}
}
