% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rds.R
\name{save_rds}
\alias{save_rds}
\title{save_rds}
\usage{
save_rds(object, filename = NULL, path = NULL, overwrite = FALSE,
  ...)
}
\arguments{
\item{object}{R object to serialize}

\item{filename}{Name of the file where the R object is saved. The file extension must be .rds}

\item{path}{Path to where the R object is saved}

\item{overwrite}{Overwrite if the file alreay exists}

\item{...}{Options passed to saveRDS}
}
\description{
save_rds
}
\details{
This function checks if the file aready exists before saving it.
See ?base::saveRDS() for more information
}
