% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rds.R
\name{save_rds}
\alias{save_rds}
\title{save_rds}
\usage{
save_rds(
  object,
  filename = NULL,
  path = NULL,
  overwrite = FALSE,
  rm_object = FALSE,
  ...
)
}
\arguments{
\item{object}{R object to serialize.}

\item{filename}{Name of the file where the R object is saved (file extension must be .rds).}

\item{path}{Path to where the R object is saved.}

\item{overwrite}{If true, existing file is overwritten.}

\item{rm_object}{If true, the object is removed form the environment.}

\item{...}{Options passed to saveRDS.}
}
\description{
Save rds
}
\details{
This function checks if the file aready exists before saving it. If the file aready
exists and `overwrite = FALSE` (default), the file is not saved. The filename and path
is combined using `file.path()`. See `?base::saveRDS()`` for more information.

It viss possible to set `path = ""` and set the entire path including the name as
the `filename` argument.
}
\examples{
\dontrun{
set.seed(42)
x <- runif(n = 100)
save_rds(x, filename = "random_numbers.rds")
}

}
\seealso{
\code{\link{saveRDS}}
}
