% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_memory_use.R
\name{show_memory_use}
\alias{show_memory_use}
\title{show_memory_use}
\usage{
show_memory_use(sort = "size", units = "Kb", decreasing = FALSE)
}
\arguments{
\item{sort}{Sort the resulting dataframe either by "size", "alphabetical" or "class".}

\item{units}{Unit used to format memory usage. For more details, see \code{?object.size}.}

\item{decreasing}{Logical if results should be sorted in decreasing order.}
}
\description{
Show memory use
}
\details{
Functions returns the name, class and size of all objects in the environment. The
resulting data frame can be sorted by different arguments.
}
\examples{
vec <- 1:100
df <- data.frame(1:5000)
mat <- matrix(c(1:100), ncol = 10)

show_memory_use()

}
\seealso{
\code{\link{object.size}}
}
