\name{support-package}
\alias{support-package}
\alias{support}
\docType{package}
\title{Support Points}
\description{
The 'support' package provides functions for generating support points.
}
\details{
\tabular{ll}{
Package: \tab support\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-08-19\cr
License: \tab GPL (>= 2)\cr
}
The 'support' package provides the function \code{sp()} for generating the support points proposed in Mak and Joseph (2017) <arXiv:1609.01811>. Support points are representative points of a possibly non-uniform distribution, and can be used as optimal sampling or integration points for a distribution of choice. The provided function \code{sp()} can be used to generate support points for standard distributions or for reducing big data (e.g., from Markov-chain Monte Carlo methods). This work was supported by USARO grant W911NF-14-1-0024 and NSF DMS grant 1712642.}
\author{
Simon Mak

Maintainer: Simon Mak <smak6@gatech.edu>
}
\references{
Mak, S. and Joseph, V. R. (2017). Support points. \emph{Annals of Statistics}. To appear.
}
\keyword{ package }

