\name{support-package}
\alias{support-package}
\alias{support}
\docType{package}
\title{Support Points}
\description{
The 'support' package provides functions for computing support points.
}
\details{
\tabular{ll}{
Package: \tab support\cr
Type: \tab Package\cr
Version: \tab 0.1.4\cr
Date: \tab 2019-07-15\cr
License: \tab GPL (>= 2)\cr
}
The 'support' package provides the functions \code{sp()} and \code{sp_seq()} for computing the support points in Mak and Joseph (2018) <DOI:10.1214/17-AOS1629>. Support points can be used as a representative sample of a desired distribution, or a representative reduction of a big dataset (e.g., an "optimal" thinning of Markov-chain Monte Carlo sample chains). This work was supported by USARO grant W911NF-14-1-0024 and NSF DMS grant 1712642.}
\author{
Simon Mak

Maintainer: Simon Mak <sm769@duke.edu>
}
\references{
Mak, S. and Joseph, V. R. (2018). Support points. \emph{Annals of Statistics}, 46(6A):2562-2592.
}
\keyword{ package }

