% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppoRt-package.R
\docType{package}
\name{suppoRt}
\alias{suppoRt}
\alias{suppoRt-package}
\title{suppoRt}
\description{
Collection of helper functions
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mhesselbarth/suppoRt}
  \item Report bugs at \url{https://github.com/mhesselbarth/suppoRt/issues}
}

}
\author{
\strong{Maintainer}: Maximilian H.K. Hesselbarth \email{maximililan.hesselbarth@uni-goettingen.de} (0000-0003-1125-9918)

}
